asoc: Add xt logging disable mixer control
Add xt logging disable mixer control in order
to differentiate whether excursion logging data
needs to be sent or not for the use case.
This is required for the cases where same backend
is used for multiple devices and xt logging data
needs to be logged only for a particular device.

Change-Id: I2cec67ddc18cbd63c5bc678eba77ea96b9a13963
Signed-off-by: Aditya Bavanari <abavanar@codeaurora.org>