qcacld-3.0: Add VHT capability in the wiphy for 2.4ghz
Currently the driver is not included the VHT capability
on 2.4ghz band if gEnableVhtFor24GHzBand ini is enabled.
As a result the driver is not advertising the VHT support
on 2.4ghz band in wiphy per band info even gEnableVhtFor24GHzBand
ini is enabled.

To resolve this, add VHT capability in the wiphy for 2.4ghz

Change-Id: If36b0ad9bf546e54542485637a99c1ee82b1850e
CRs-Fixed: 2495337
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>