qca-wifi: Add dfs_set_cfreq2 flag in multivdev_restart channel params
While switching to 36 HT160 as part of multivdev restart, the channel
params to be sent to FW are only updated with primary segment dfs set flag.

Consider if the primary channel is 36 (non DFS), in which case, the
primary HT80 segment is non DFS, but secondary HT80 segment is DFS. This
will be indicated to FW by sending the flags with dfs_set as FALSE and
dfs_set_cfreq2 as TRUE. But since dfs_set_cfreq2 is not updated incase of
multivdev restart, FW does not detect radar in the secondary detector.

Add proper dfs_set_cfreq2 flag in multivdev_restart channel params.

Change-Id: I875534c3f61b1d9a36efe1ec63505332a59a4dd1
CRs-Fixed: 2496923
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>