qca-wifi: Reset the picked agile channel if it is an excluded channel
While selecting a channel for aDFS, the current channel(s) are skipped
for preCAC (since they are CACed by the primary detector).

If no other channels are found and if the current channel
is the last in the preCAC list, aDFS is started on this channel.
This is because, the channel variable defined in the API locally
is not reset if it is an invalid channel.

Reset the agile channel at the end of the loop if it is one of the
excluded (current) channel.

Change-Id: I48cbfc3c511c05bed86b60eb9e2ca22f9b8e0362
CRs-Fixed: 2515324
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>