qcacld-3.0: Update SAE status to SME in SAP mode as well
A check has been introduced while implementing SAE roaming to send
the SAE status to SME only when the STA is not in joined state.
But this doesn't handle the SAP case where the status is expected
to be updated always.

Update the status to SME in below cases
1. SAP mode: Always
2. STA mode: When the device is not in joined state

If the device is in joined state, send the status to WMA which
is meant for roaming.

Change-Id: I20c134a3eea7c65db1fe1950c02b8b9e1ee3a34d
CRs-Fixed: 2524973
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>