qcacld-3.0: Fix peer phymode for STA while updating opmode
In STA use iface->chanmode which is updated during vdev restart to
set the bss peer channel width.

Also remove the API to set phymode during this as the phymode is
already set during vdev start or restart, so no need to set the
phymode again.

So just check if the new BW is valid for the current phymode
and if so set the BW for the peer.

Change-Id: Ic258703acbab0a97629ff4e1945f6729b3fe517f
CRs-Fixed: 2535576
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>